新增gbk2312编码以及长按12秒复位功能，结构修改了打字机的表示

+++ 代码/User/ch32v00x_it.h
@@ -42,7 +42,7 @@ typedef struct
   KeyState state;          // 当前按键状态
   KeyEvent event;          // 当前按键事件
   u8 debounce_counter;     // 去抖动计数器
-  uint16_t LongKeyCounter; // 长按计数器
+  uint32_t LongKeyCounter; // 长按计数器
   u8 enable;               // 使能
 
 } Key;
